A start to making subqueries more expressive.
This adds the notion of an underlyig model as well as model constraints to subqueries. This will allow for subqueries to be based on a model other than the primary model of the outer query. Also, adding the notion of constraints to sub-queries will allow for JOIN "ON" constraints vs. having to use WHERE constraints for everything. I am hoping that separating these notions more fully will allor for features like predicate push-down.